requirements.txt
Django==2.1.8
django-redis==4.10.0
django-stubs==0.12.1
djangorestframework==3.9.2
djangorestframework-stubs==0.4.2
mypy==0.701
mypy-extensions==0.4.1
mysqlclient==1.4.2.post1
Pillow==5.4.1
pytz==2018.9
redis==3.2.1
typed-ast==1.3.4
typing-extensions==3.7.2
uWSGI==2.0.18
